Kill stickiness.py
