+++ ee/clickhouse/queries/stickiness/clickhouse_stickiness.py
@@ -1,5 +1,7 @@
+import copy
+import urllib.parse
 from datetime import datetime
-from typing import Any, Dict, Tuple
+from typing import Any, Dict, List
 
 from django.conf import settings
 from django.db.models.expressions import F
@@ -12,14 +14,28 @@ from ee.clickhouse.queries.person_distinct_id_query import get_team_distinct_ids
 from ee.clickhouse.queries.stickiness.stickiness_actors import ClickhouseStickinessActors
 from ee.clickhouse.queries.stickiness.stickiness_event_query import StickinessEventsQuery
 from ee.clickhouse.sql.person import GET_LATEST_PERSON_SQL, INSERT_COHORT_ALL_PEOPLE_SQL, PERSON_STATIC_COHORT_TABLE
+from posthog.constants import TREND_FILTER_TYPE_ACTIONS
+from posthog.models.action import Action
 from posthog.models.cohort import Cohort
 from posthog.models.entity import Entity
 from posthog.models.filters.stickiness_filter import StickinessFilter
 from posthog.models.team import Team
-from posthog.queries.stickiness import Stickiness
+from posthog.queries.base import handle_compare
+from posthog.utils import encode_get_request_params
 
 
-class ClickhouseStickiness(Stickiness):
+class ClickhouseStickiness:
+    def run(self, filter: StickinessFilter, team: Team, *args, **kwargs) -> List[Dict[str, Any]]:
+
+        response = []
+        for entity in filter.entities:
+            if entity.type == TREND_FILTER_TYPE_ACTIONS:
+                entity.name = Action.objects.only("name").get(team=team, pk=entity.id).name
+
+            entity_resp = handle_compare(filter=filter, func=self._serialize_entity, team=team, entity=entity)
+            response.extend(entity_resp)
+        return response
+
     def stickiness(self, entity: Entity, filter: StickinessFilter, team_id: int) -> Dict[str, Any]:
         events_query, event_params = StickinessEventsQuery(entity, filter, team_id).get_query()
 
@@ -33,9 +49,62 @@ class ClickhouseStickiness(Stickiness):
         counts = sync_execute(query, {**event_params, "num_intervals": filter.total_intervals})
         return self.process_result(counts, filter, entity)
 
-    def _retrieve_people(self, target_entity: Entity, filter: StickinessFilter, team: Team, request: Request):
+    def people(self, target_entity: Entity, filter: StickinessFilter, team: Team, request, *args, **kwargs):
         _, serialized_actors = ClickhouseStickinessActors(entity=target_entity, filter=filter, team=team).get_actors()
         return serialized_actors
+
+    def process_result(self, counts: List, filter: StickinessFilter, entity: Entity) -> Dict[str, Any]:
+        response: Dict[int, int] = {}
+        for result in counts:
+            response[result[1]] = result[0]
+
+        labels = []
+        data = []
+        for day in range(1, filter.total_intervals):
+            label = "{} {}{}".format(day, filter.interval, "s" if day > 1 else "")
+            labels.append(label)
+            data.append(response[day] if day in response else 0)
+        filter_params = filter.to_params()
+
+        return {
+            "labels": labels,
+            "days": [day for day in range(1, filter.total_intervals)],
+            "data": data,
+            "count": sum(data),
+            "filter": filter_params,
+            "persons_urls": self._get_persons_url(filter, entity),
+        }
+
+    def _serialize_entity(self, entity: Entity, filter: StickinessFilter, team_id: int) -> List[Dict[str, Any]]:
+        serialized: Dict[str, Any] = {
+            "action": entity.to_dict(),
+            "label": entity.name,
+            "count": 0,
+            "data": [],
+            "labels": [],
+            "days": [],
+        }
+        response = []
+        new_dict = copy.deepcopy(serialized)
+        new_dict.update(self.stickiness(entity=entity, filter=filter, team_id=team_id))
+        response.append(new_dict)
+        return response
+
+    def _get_persons_url(self, filter: StickinessFilter, entity: Entity) -> List[Dict[str, Any]]:
+        persons_url = []
+        for interval_idx in range(1, filter.total_intervals):
+            filter_params = filter.to_params()
+            extra_params = {
+                "stickiness_days": interval_idx,
+                "entity_id": entity.id,
+                "entity_type": entity.type,
+                "entity_math": entity.math,
+            }
+            parsed_params: Dict[str, str] = encode_get_request_params({**filter_params, **extra_params})
+            persons_url.append(
+                {"filter": extra_params, "url": f"api/person/stickiness/?{urllib.parse.urlencode(parsed_params)}",}
+            )
+        return persons_url
 
 
 def insert_stickiness_people_into_cohort(cohort: Cohort, target_entity: Entity, filter: StickinessFilter) -> None:
